+++ lib/authentication.js
@@ -1,8 +1,10 @@
 const fetch = require('node-fetch');
+const dotenv = require('dotenv');
+const util = require('util');
 const api = require('../helpers/api');
 const jwt = require('../helpers/jwt');
-const util = require('util');
 
+globalizeAppId();
 
 function WrongAudienceError(message) {
   this.message = message;
@@ -72,3 +74,12 @@ exports.verifyAudience = function (token, audiences) {
   });
   return !!intersection.length;
 };
+
+function globalizeAppId() {
+  dotenv.config();
+  global.APP_ID = process.env.APP_ID;
+  if (!global.APP_ID) {
+    const msg = 'App id not available, did you deploy using DroneDeploy-Cli?';
+    throw new Error(msg);
+  }
+}
